Trial multiprocessing to speed up variant table build
Since our patch to speed up Asklepian by swapping fastafunk for
gofasta several months ago, the speed gained has slowly but surely
been ebbed away with the linear scaling of our ever-growing data set.

In lieu of an easy way to implement deltas (both internally, but
especially via EDGE), I have liberally applied CPU to the affected
area. Using multiprocessing, the variant table builder will now
instantiate multiple processes, each with a file handle to the MSA and
an assigned offset to the MSA. The processes will each compute the
variants for their assigned block of the MSA using the exact same
algorithm as the existing variant table. Processes will emit a single
string containing their result to a shared multiprocessing Queue that
is read by a single writer process.

While performance may seem lacklustre at first, the multiprocessing
Queue is saturated early and obtaining the Queue lock becomes the
limiting factor to writing the variant table. During this time, the
Queue feeder threads will grow quite large as the variant workers buffer
their output (without interrupting calculation of variants) in RAM.

As the Queue.put calls ease due to variant worker processes completing
their business and draining the buffered feeders; the writer process
will ramp up to call Queue.get at the speed of ceph -- yielding a
significant performance increase over the previous linear table building.

On the Elan 20220227 data set, the v2 variant table builder used 8
processes to construct the table in 24 minutes. The v1 variant table
builder constructed the table in 270 minutes.

The v2 variant table will have an unpredictable sample order (in that
each process is free to emit samples as fast as they can, and locking
will be non-deterministic) but sorting the CSV lines showed that the
contents of the v2 variant table matched that of the v1 table.

The number of processes is controlled with `ASKLEPIAN_VARIANT_THREADS`
and can be increased if the data set continues to grow beyond our
expectations.

With a clear, significant performance gain and a reasonable
demonstration of determinism, it is recommended that after a trial run
that the v2 table supercede the v1 variant table as soon as practical.